move helper function to separate library
